Nicholas (#8)
* Added Timer Labels

* player

* moved create audio

* SeekBar Functionality

I set up the seekBar and implemented a separate function called setup
music that can setup the seeker and music player for any song Id.

* Implemented Time Labels and progress time

Implemented Time Labels and progress time. I am using a while statement
that checks for whether or not media player is null and whether or not
current position is less than total.